+++ src/app/(route)/customer/[customerCategory]/write/page.tsx
@@ -9,7 +9,7 @@ import { useEditStore } from "@/utils/Store";
 import axios from "axios";
 import getUpload from "@/_hooks/getUpload";
 import usePutPost from "@/_hooks/fetcher/board/usePutPost";
-import CustomerTiptap from "../../_components/CustomerTiptap";
+import CustomerTiptap from "../../_components/ui/CustomerTiptap";
 import { cn } from "@/utils";
 import { useToast } from "@/_hooks/useToast";
 import { useAdminRole } from "../../_utils/adminChecker";
