+++ src/app/(route)/customer/_components/common/TermsLayout.tsx
