+++ src/app/(route)/customer/feedback/feedback-info/[feedbackId]/_components/FeedbackInfo.tsx
@@ -23,12 +23,12 @@ import Image from "next/image";
 import RecommendButton from "@/app/(route)/(community)/_components/RecommendButton";
 import PostAction from "@/app/(route)/(community)/_components/PostAction";
 import PostNavigation from "@/app/(route)/(community)/_components/PostNavigation";
-import CustomerTalkToolbar from "@/app/(route)/customer/_components/CustomerTalkToolbar";
-import FeedbackItemSkeleton from "@/app/(route)/customer/_components/FeedbackItemSkeleton";
-import NoticeItem from "@/app/(route)/customer/_components/NoticeItem";
-import EmptyItem from "@/app/(route)/customer/_components/EmptyItem";
+import CustomerTalkToolbar from "@/app/(route)/customer/_components/ui/CustomerTalkToolbar";
+import FeedbackItemSkeleton from "@/app/(route)/customer/feedback/_components/status/FeedbackItemSkeleton";
+import NoticeItem from "@/app/(route)/customer/notice/_components/items/NoticeItem";
+import EmptyItem from "@/app/(route)/customer/_components/common/EmptyItem";
 import { FeedbackContentType } from "@/app/(route)/customer/_types/FeedbackItemType";
-import FeedbackItem from "@/app/(route)/customer/_components/FeedbackItem";
+import FeedbackItem from "@/app/(route)/customer/feedback/_components/items/FeedbackItem";
 import SignInModalPopUp from "@/app/_components/SignInModalPopUp";
 import BoardComment from "@/app/(route)/(community)/_components/BoardComment";
 import { CommentItem } from "@/_types/comment";
