+++ src/app/(route)/customer/notice/_components/items/NoticeItem.tsx
