+++ src/app/(route)/customer/notice/_components/status/NoticeItemSkeleton.tsx
