+++ src/app/(route)/customer/notice/notice-info/[noticeInfoId]/_components/NoticeInfo.tsx
@@ -8,10 +8,10 @@ import { useParams, useRouter, useSearchParams } from "next/navigation";
 import React, { Suspense } from "react";
 import NoticeInfoItemSkeleton from "./NoticeInfoItemSkeleton";
 import NoticeInfoItem from "./NoticeInfoItem";
-import CustomerTalkToolbar from "@/app/(route)/customer/_components/CustomerTalkToolbar";
-import NoticeItemSkeleton from "@/app/(route)/customer/_components/NoticeItemSkeleton";
-import EmptyItem from "@/app/(route)/customer/_components/EmptyItem";
-import NoticeItem from "@/app/(route)/customer/_components/NoticeItem";
+import CustomerTalkToolbar from "@/app/(route)/customer/_components/ui/CustomerTalkToolbar";
+import NoticeItemSkeleton from "@/app/(route)/customer/notice/_components/status/NoticeItemSkeleton";
+import EmptyItem from "@/app/(route)/customer/_components/common/EmptyItem";
+import NoticeItem from "@/app/(route)/customer/notice/_components/items/NoticeItem";
 import { NoticeContentType } from "@/app/(route)/customer/_types/NoticeItemType";
 import { cn } from "@/utils";
 import Pagination from "@/app/(route)/mypage/_components/Pagination";
