+++ src/app/(route)/news/[newsCategoryType]/news-detail/[id]/_components/NewsRecommend.tsx
@@ -1,5 +1,5 @@
 import useSortedNewsDataList from "@/_hooks/fetcher/news/useSortedNewsDataList";
-import EmptyItem from "@/app/(route)/customer/_components/EmptyItem";
+import EmptyItem from "@/app/(route)/customer/_components/common/EmptyItem";
 import Pagination from "@/app/(route)/mypage/_components/Pagination";
 import NewsPostItem from "@/app/(route)/news/_components/NewsPostItem";
 import NewsPostItemSkeleton from "@/app/(route)/news/_components/NewsPostItemSkeleton";
